Correct timer 0 high and SPI0 settings, so they'll work with the gLCD. Fix pin assignments for GLCD's A0 and RST.
Finish coding the API's UpdateFrameBuffer function (not tested).
Enable GLCD in main.

In nec_ir.c, remember to clear Timer interrupt in the ISR. Variables set in the ISR need to be volatile.  Make sure timer count is preset to the right re-load value before we restart the timer for each bit.
In nec_ir.h, fix endian-ness of the packet bytes in the structure.