Support both IR formats. Support Apple Remote format. Rename some functions so we can tell the difference between the RC5 and the NEC IR formats.
Change some const code to #define to save code space.

Need to work up the command parser in the main loop. I set up a list of commands which each individual receiver will return. I have to make button press on and off obvious and then revamp the parser to work with that new command.