hardware.h: fix the definition of NEC_EXTINT_LOW and the SBIT assignment for the encoder switch.
ircmdhandler.c: volume up and down are the same as plus and minus. Repeat implemented for volume up and down. Disallow repeats for mute toggle.

lcd.c: make sure to #ifdef out the prototype for isLcdBusy().

main.c: #ifdef INCL_BARGRAPH_TEST is now #if INCL_BARGRAPH_TEST == 1, which is the right way to do this (the macro is defined always, and has a value 1 or 0).

nec_ir.c: increase timing fudge factor.

quaddecoder.c: fix debounce decrement so it actually goes to 0. 
In timer ISR, only grab pins if the main loop has checked the previous value.