in midi_uart: transmit FIFO almost full threshold is actually half full, to ensure there is plenty of space in the FIFO for USB packets. Add a function isRoomInFifo() so main loop code can test for that condition.
in main(): test for roomInFifo is done outside of message popping and parsing, once per run through the main loop. If there is room enough to not overrun the serial transmitter FIFO, and if the endpoint isn’t already busy, we call USBD_Read() to prepare for the next USB bulk packet. If we don’t call that function, and packets arrive, those packets will fill the endpoint FIFO, at which time the SIE will NAK subsequent transactions. Once the serial port has popped and sent enough bytes for there to be room in its FIFO, we call USBD_Read() to let more packets in.